third_party/WebKit/LayoutTests/user-activation-v2/full-screen-gamepad.html

<body>
<script src="../fullscreen/full-screen-test.js"></script>
<span></span>
<script>
    var span = document.getElementsByTagName('span')[0];

    var documentEnteredFullscreen = function() {
        testExpected("document.fullscreenElement", document.documentElement);
        span.requestFullscreen();
        waitForEvent(document, 'fullscreenchange', spanEnteredFullscreen, false, true);
    };

    var spanEnteredFullscreen = function() {
        testExpected("document.fullscreenElement", span);
        waitForEvent(document, 'fullscreenchange', spanExited, false, true);
        document.exitFullscreen();
    };

    var spanExited = function() {
        testExpected("document.fullscreenElement", document.documentElement);
        waitForEvent(document, 'fullscreenchange', documentExited, false, true);
        document.exitFullscreen();
    };

    var documentExited = function() {
        testExpected("document.fullscreenElement", undefined);
        endTest();
    };

    testExpected("internals.runtimeFlags.userActivationV2Enabled",
                 internals.runtimeFlags.userActivationV2Enabled);
    if (!internals.runtimeFlags.userActivationV2Enabled) {
        console.log("This test requires User Activation V2.");
        endTest();
    } else {
        if (window.gamepadController) {
            // Simulate pressing a button on a connected gamepad. A gamepad
            // button press should be treated as a user activation for the
            // fullscreen API.
            gamepadController.connect(0);
            gamepadController.setId(0, "FullScream Maximizer");
            gamepadController.setButtonCount(0, 1);
            gamepadController.setButtonData(0, 0, 1);
            navigator.getGamepads();
        }
        document.documentElement.requestFullscreen();
        waitForEvent(document, 'fullscreenchange', documentEnteredFullscreen, false, true);
    }
</script>
