third_party/WebKit/LayoutTests/user-activation-v2/full-screen-gamepad-expected.txt

CONSOLE MESSAGE: line 33: This test requires User Activation V2.
EXPECTED (internals.runtimeFlags.userActivationV2Enabled == 'false') OK
END OF TEST
